+++ bundle/src/test/java/com/adobe/acs/commons/wcm/impl/RobotsServletTest.java
@@ -67,6 +67,16 @@ public class RobotsServletTest {
     }
 
     @Test
+    public void testCrawlerDelay() throws IOException, ServletException {
+        Map<String, Integer> props = new HashMap<>();
+        props.put("crawl.delay", 60);
+        RobotsServlet robotsServlet = context.registerInjectActivateService(new RobotsServlet(), props);
+        robotsServlet.doGet(request, response);
+        assertEquals("servlet returned an error", 200, response.getStatus());
+        assertResponse("RobotsServlet_testCrawlerDelay.txt", response);
+    }
+
+    @Test
     public void testWriteFromPageProperty() throws IOException, ServletException {
         Map<String, Object> props = new HashMap<>();
         props.put("sling.servlet.resourceTypes", "geometrixx/components/structure/page");